+++ docs/source/changes.md
@@ -187,6 +187,12 @@ authors:
   ({{ gh(606) }} by {{ authors.larsoner }})
 - Add SSP joint plots to generated reports
 - ({{ gh(614) }} by {{ authors.larsoner }})
+- It is now possible to specify baseline and cropping parameters for
+  plotting time-frequency results via
+  [`time_frequency_baseline`][mne_bids_pipeline.config.time_frequency_baseline],
+  [`time_frequency_baseline_mode`][mne_bids_pipeline.config.time_frequency_baseline_mode]
+  and [`time_frequency_crop`][mne_bids_pipeline.config.time_frequency_crop].
+  ({{ gh(641) }} by {{ authors.agramfort }})
 - Add time-frequency decoding based on [common spatial patterns](https://mne.tools/stable/auto_examples/decoding/decoding_csp_timefreq.html)
    (CSP).
    ({{ gh(625) }} by {{ authors.crsegerie }}, {{ authors.agramfort }},